app/pyproject.toml
[project]
name = "vertretungsplan-app"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "markupsafe>=3.0.2",
    "psycopg2-binary>=2.9.10",
]
